relax audio_filepath requirement in contours; refactor PeakStreamHelper into utils
